Update AIChat components and dependencies for chat functionality
