+++ components/AIChat.tsx
@@ -4,16 +4,10 @@ import { Bot, Trash, SendHorizontal, XCircle } from 'lucide-react'
 import React, { useEffect, useRef, useState } from 'react'
 import { cn } from '@/lib/utils'
 import { toast } from 'react-toastify'
-import { Message, useChat } from 'ai/react'
-import Link from 'next/link'
-import ReactMarkdown from 'react-markdown'
+import { useChat } from 'ai/react'
+import AIChatMessage from '@/components/AIChatMessage'
 
 const AIChat = () => {
-  const [open, setOpen] = useState(false)
-  const [sourcesForMessages, setSourcesForMessages] = useState<
-    Record<string, unknown>
-  >({})
-
   const {
     messages,
     input,
@@ -46,6 +40,11 @@ const AIChat = () => {
 
   const inputRef = useRef<HTMLInputElement>(null)
   const scrollRef = useRef<HTMLDivElement>(null)
+
+  const [open, setOpen] = useState(false)
+  const [sourcesForMessages, setSourcesForMessages] = useState<
+    Record<string, unknown>
+  >({})
 
   useEffect(() => {
     if (scrollRef.current) {
@@ -137,53 +136,6 @@ const AIChat = () => {
             </button>
           </form>
         </div>
-      </div>
-    </div>
-  )
-}
-
-interface AIChatMessageProps {
-  message: Message
-}
-
-const AIChatMessage = ({ message: { role, content } }: AIChatMessageProps) => {
-  const isAiMessage = role === 'assistant'
-
-  return (
-    <div
-      className={cn(
-        'mb-3 flex items-center',
-        isAiMessage ? 'me-5 justify-start' : 'ms-5 justify-end',
-      )}
-    >
-      {isAiMessage && <Bot className="mr-2 flex-none" />}
-      <div
-        className={cn(
-          'rounded-md border px-3 py-2',
-          isAiMessage ? 'bg-background' : 'bg-foreground text-background',
-        )}
-      >
-        <ReactMarkdown
-          components={{
-            a: ({ ...props }) => (
-              <Link
-                {...props}
-                href={props.href ?? ''}
-                className="text-primary hover:underline"
-              />
-            ),
-            p: ({ ...props }) => <p {...props} className="mt-3 first:mt-0" />,
-            ul: ({ ...props }) => (
-              <ul
-                {...props}
-                className="mt-3 list-inside list-disc first:mt-0"
-              />
-            ),
-            li: ({ ...props }) => <li {...props} className="mt-1" />,
-          }}
-        >
-          {content}
-        </ReactMarkdown>
       </div>
     </div>
   )
