+++ data/index.ts
@@ -1,3 +1,9 @@
+import { Message } from 'ai'
+
+export interface AIChatMessageProps {
+  message: Message
+}
+
 export interface Dictionary {
   page: {
     experience: string
